+++ Cyan-Stars/Assets/Scripts/Libraries/MarkdownRenderer/Runtime/Extensions/AtParagraph/AtParagraphInline.cs
@@ -1,0 +1,16 @@
+using Markdig.Syntax.Inlines;
+
+namespace CyanStars.MarkdownRenderer.Extensions.AtParagraph
+{
+    public sealed class AtParagraphInline : LinkInline
+    {
+        public AtParagraphInline(string paragraph)
+        {
+            Paragraph = paragraph;
+            Label = "@" + paragraph;
+            Url = string.Empty;
+        }
+
+        public string Paragraph { get; }
+    }
+}
